+++ src/chunk.c
@@ -93,6 +93,7 @@ const char *opcode_nombre(OpCode op) {
         case OP_ITER_SIGUIENTE:  return "OP_ITER_SIGUIENTE";
         case OP_FORMATO_F:       return "OP_FORMATO_F";
         case OP_ASEGURAR_CADENA: return "OP_ASEGURAR_CADENA";
+        case OP_REPR:            return "OP_REPR";
         case OP_LONGITUD:        return "OP_LONGITUD";
         case OP_ES_TUPLA:        return "OP_ES_TUPLA";
         case OP_ES_LISTA:        return "OP_ES_LISTA";